+++ MonkeyDelivery/Src/Control/UI/Buttons/Back.h
@@ -6,7 +6,7 @@ class Game;
 class Back : public Button {
 
 public:
-	Back(int x, int y,int w , int h,Game* game);
+	Back(int x, int y,double w ,double h,Game* game);
 	~Back() {};
 	void onCursorCollision() override;
 };
