+++ MonkeyDelivery/Src/Control/UI/Buttons/Back1.cpp
@@ -4,7 +4,7 @@
 #include "../../States/PlayingState.h"
 
 
-Back1::Back1(int x, int y, int w, int h, Game* game):Button(game,w,h,x,y){
+Back1::Back1(int x, int y, double w, double h, Game* game):Button(game,w,h,x,y){
 	setTexture(button_Back);
 }
