+++ MonkeyDelivery/Src/Control/UI/Buttons/Back1.h
@@ -6,7 +6,7 @@ class Game;
 class Back1 : public Button {
 
 public:
-	Back1(int x, int y,int w , int h,Game* game);
+	Back1(int x, int y,double w ,double h,Game* game);
 	~Back1() {};
 	void onCursorCollision() override;
 };
