+++ MonkeyDelivery/Src/Control/UI/Buttons/Continue.cpp
@@ -3,7 +3,7 @@
 #include "../../../Logic/Game.h"
 #include "../../States/PlayingState.h"
 
-Continue::Continue(int x, int y, int w, int h, Game* game) :Button(game, w, h, x, y)
+Continue::Continue(int x, int y, double w, double h, Game* game) :Button(game, w, h, x, y)
 {
 	setTexture(button_Continue);
 }
