+++ MonkeyDelivery/Src/Control/UI/Buttons/ExitButton.h
@@ -3,6 +3,6 @@
 
 class ExitButton :public Button {
 public:
-	ExitButton(int x, int y, int w, int h, Game* game);
+	ExitButton(int x, int y, double w, double h, Game* game);
 	void onCursorCollision() override;
 };
