+++ MonkeyDelivery/Src/Control/UI/Buttons/Options.cpp
@@ -2,7 +2,7 @@
 #include "../../../Logic/Game.h"
 #include "../../States/OptionsState.h"
 
-Options::Options(int x, int y, int w, int h, Game* game, int num) :Button(game, w, h, x, y){
+Options::Options(int x, int y, double w, double h, Game* game, int num) :Button(game, w, h, x, y){
 	setTexture(button_Main_Options);
 	num_ = num;
 }
