+++ MonkeyDelivery/Src/Control/UI/Buttons/Options.h
@@ -8,7 +8,7 @@ class Options : public Button {
 	int num_;
 
 public:
-	Options(int x, int y, int w, int h, Game* game, int num);
+	Options(int x, int y, double w, double h, Game* game, int num);
 	~Options() {};
 	void onCursorCollision() override;
 };
