+++ MonkeyDelivery/Src/Control/UI/Buttons/Pause.h
@@ -6,7 +6,7 @@ class Game;
 class Pause : public Button {
 
 public:
-	Pause(int x, int y,int w , int h,Game* game);
+	Pause(int x, int y, double w , double h,Game* game);
 	~Pause() {};
 	void onCursorCollision() override;
 };
