+++ MonkeyDelivery/Src/Control/UI/Buttons/Start.h
@@ -5,7 +5,7 @@ class Game;
 
 class Start : public Button {
 public:
-	Start(int x, int y,int w , int h,Game* game);
+	Start(int x, int y,double w , double h,Game* game);
 	~Start() {};
 	void onCursorCollision() override;
 };
